backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "arxiv>=2.3.1",
    "auth>=1.2.1",
    "bcrypt==4.3.0",
    "braintrust>=0.3.13",
    "braintrust-langchain>=0.1.5",
    "duckduckgo-search>=8.1.1",
    "fastapi[standard]>=0.124.2",
    "google-search-results>=2.4.2",
    "jinja2>=3.1.6",
    "jsonschema>=4.25.1",
    "langchain>=1.1.3",
    "langchain-cli>=0.0.37",
    "langchain-community>=0.4.1",
    "langchain-core>=1.1.3",
    "langchain-experimental>=0.4.1",
    "langchain-openai>=1.1.2",
    "langchain-tools>=0.1.34",
    "langgraph>=1.0.4",
    "langgraph-checkpoint-postgres>=3.0.2",
    "newsapi-python>=0.2.7",
    "numpy>=2.3.5",
    "openai>=2.11.0",
    "pandas>=2.3.3",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-ai>=1.31.0",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "rich>=14.2.0",
    "serpapi>=0.1.5",
    "sqlalchemy>=2.0.45",
    "tenacity>=9.1.2",
    "tqdm>=4.67.1",
    "uvicorn>=0.38.0",
    "wikipedia>=1.4.0",
]
